Updated data access layer to new design
Rather than injecting repositories into each service, there is now an IDatabase object that manages the repositories and the implements the Unit of Work pattern.  This is what is injected into Services.  As such, an IFoodTruckDatabase interface was added, as well as a FoodTruckDatabase object as the concrete implementation.  Services were updated for this new approach.

Some FoodTruck classes had to be updated with Locality when a new Food Truck was created in order to test adding a FoodTruck using these revised data access classes.  This was missed on a previous PR.